Move get_array_in_code_units function
