+++ plonk/snap/utils.py
@@ -4,6 +4,9 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from numpy import ndarray
+
+from .._units import Quantity
 from .._units import units as plonk_units
 
 if TYPE_CHECKING:
@@ -27,3 +30,24 @@ def gravitational_constant_in_code_units(snap: SnapLike) -> float:
     G_units = snap.units['length'] ** 3 / snap.units['mass'] / snap.units['time'] ** 2
     G = (G / G_units).to_base_units().magnitude
     return G
+
+
+def get_array_in_code_units(snap: SnapLike, name: str) -> ndarray:
+    """Get array in code units.
+
+    Parameters
+    ----------
+    snap
+        The Snap or SubSnap.
+    name
+        The array name.
+
+    Returns
+    -------
+    ndarray
+        The array on the particles in code units.
+    """
+    arr = snap[name]
+    if isinstance(arr, Quantity):
+        return (arr / snap.get_array_unit(name)).magnitude
+    return arr
